feat: create socket.io connection from client to server and vice versa
