+++ index.html
@@ -6,8 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
   </head>
+  <script src="/socket.io/socket.io.js"></script>
+  <script type="module">
+    import { io } from "socket.io-client";
+    const socket = io();
+  </script>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/app/main.tsx"></script>
+    <script type="module" src="src/app/main.tsx"></script>
   </body>
 </html>
